refactor: Extract ChartMoreButton
