+++ app/components/chart-preview.tsx
@@ -6,8 +6,8 @@ import {
   useDraggable,
   useDroppable,
 } from "@dnd-kit/core";
-import { Trans, t } from "@lingui/macro";
-import { Box, IconButton, useEventCallback } from "@mui/material";
+import { Trans } from "@lingui/macro";
+import { Box } from "@mui/material";
 import { makeStyles } from "@mui/styles";
 import Head from "next/head";
 import React, {
@@ -20,7 +20,6 @@ import React, {
 
 import { DataSetTable } from "@/browse/datatable";
 import { LoadingStateProvider } from "@/charts/shared/chart-loading-state";
-import { ArrowMenu } from "@/components/arrow-menu";
 import { ChartErrorBoundary } from "@/components/chart-error-boundary";
 import { ChartFootnotes } from "@/components/chart-footnotes";
 import {
@@ -29,7 +28,7 @@ import {
   ChartWrapperProps,
 } from "@/components/chart-panel";
 import { chartPanelLayoutGridClasses } from "@/components/chart-panel-layout-grid";
-import { ChartControls } from "@/components/chart-shared";
+import { ChartControls, ChartMoreButton } from "@/components/chart-shared";
 import {
   ChartTablePreviewProvider,
   useChartTablePreview,
@@ -42,7 +41,6 @@ import { DragHandle } from "@/components/drag-handle";
 import Flex from "@/components/flex";
 import { Checkbox } from "@/components/form";
 import { HintYellow } from "@/components/hint";
-import { MenuActionItem } from "@/components/menu-action-item";
 import { BANNER_MARGIN_TOP } from "@/components/presence";
 import {
   ChartConfig,
@@ -59,7 +57,6 @@ import {
   useDataCubesMetadataQuery,
 } from "@/graphql/hooks";
 import { DataCubePublicationStatus } from "@/graphql/resolver-types";
-import SvgIcMore from "@/icons/components/IcMore";
 import { useLocale } from "@/locales/use-locale";
 import { InteractiveFiltersChartProvider } from "@/stores/interactive-filters";
 import { useTransitionStore } from "@/stores/transition";
@@ -231,61 +228,6 @@ const DashboardPreview = (props: DashboardPreviewProps) => {
 type CommonChartPreviewProps = ChartWrapperProps & {
   chartKey: string;
   dataSource: DataSource;
-};
-
-const ChartMoreButton = ({ chartKey }: { chartKey: string }) => {
-  const [anchor, setAnchor] = useState<HTMLElement | null>(null);
-  const handleClose = useEventCallback(() => setAnchor(null));
-  const [state, dispatch] = useConfiguratorState(hasChartConfigs);
-  return (
-    <>
-      <IconButton onClick={(ev) => setAnchor(ev.currentTarget)}>
-        <SvgIcMore />
-      </IconButton>
-      <ArrowMenu
-        open={!!anchor}
-        anchorEl={anchor}
-        onClose={handleClose}
-        anchorOrigin={{ horizontal: "center", vertical: "bottom" }}
-        transformOrigin={{ horizontal: "center", vertical: "top" }}
-      >
-        {isConfiguring(state) ? null : (
-          <MenuActionItem
-            type="button"
-            as="menuitem"
-            onClick={() => {
-              dispatch({ type: "CONFIGURE_CHART", value: { chartKey } });
-              handleClose();
-            }}
-            iconName="edit"
-            label={<Trans id="chart-controls.edit">Edit</Trans>}
-          />
-        )}
-        {state.chartConfigs.length > 1 ? (
-          <MenuActionItem
-            type="button"
-            as="menuitem"
-            color="error"
-            requireConfirmation
-            confirmationTitle={t({
-              id: "chart-controls.delete.title",
-              message: "Delete chart?",
-            })}
-            confirmationText={t({
-              id: "chart-controls.delete.confirmation",
-              message: "Are you sure you want to delete this chart?",
-            })}
-            onClick={() => {
-              dispatch({ type: "CHART_CONFIG_REMOVE", value: { chartKey } });
-              handleClose();
-            }}
-            iconName="trash"
-            label={<Trans id="chart-controls.delete">Delete</Trans>}
-          />
-        ) : null}
-      </ArrowMenu>
-    </>
-  );
 };
 
 const ReactGridChartPreview = forwardRef<
