+++ app/components/chart-shared.tsx
@@ -1,13 +1,23 @@
-import { Box } from "@mui/material";
-import { ComponentProps } from "react";
+import { Trans, t } from "@lingui/macro";
+import { Box, IconButton, useEventCallback } from "@mui/material";
+import { ComponentProps, useState } from "react";
 
 import {
   ChartDataFiltersList,
   ChartDataFiltersToggle,
   useChartDataFiltersState,
 } from "@/charts/shared/chart-data-filters";
+import { ArrowMenu } from "@/components/arrow-menu";
+import { MenuActionItem } from "@/components/menu-action-item";
 import { MetadataPanel } from "@/components/metadata-panel";
-import { ChartConfig, DataSource } from "@/configurator";
+import {
+  ChartConfig,
+  DataSource,
+  hasChartConfigs,
+  isConfiguring,
+  useConfiguratorState,
+} from "@/configurator";
+import SvgIcMore from "@/icons/components/IcMore";
 
 export const ChartControls = ({
   dataSource,
@@ -56,5 +66,60 @@ export const ChartControls = ({
         {showFilters && <ChartDataFiltersList {...chartFiltersState} />}
       </Box>
     </Box>
+  );
+};
+
+export const ChartMoreButton = ({ chartKey }: { chartKey: string }) => {
+  const [anchor, setAnchor] = useState<HTMLElement | null>(null);
+  const handleClose = useEventCallback(() => setAnchor(null));
+  const [state, dispatch] = useConfiguratorState(hasChartConfigs);
+  return (
+    <>
+      <IconButton onClick={(ev) => setAnchor(ev.currentTarget)}>
+        <SvgIcMore />
+      </IconButton>
+      <ArrowMenu
+        open={!!anchor}
+        anchorEl={anchor}
+        onClose={handleClose}
+        anchorOrigin={{ horizontal: "center", vertical: "bottom" }}
+        transformOrigin={{ horizontal: "center", vertical: "top" }}
+      >
+        {isConfiguring(state) ? null : (
+          <MenuActionItem
+            type="button"
+            as="menuitem"
+            onClick={() => {
+              dispatch({ type: "CONFIGURE_CHART", value: { chartKey } });
+              handleClose();
+            }}
+            iconName="edit"
+            label={<Trans id="chart-controls.edit">Edit</Trans>}
+          />
+        )}
+        {state.chartConfigs.length > 1 ? (
+          <MenuActionItem
+            type="button"
+            as="menuitem"
+            color="error"
+            requireConfirmation
+            confirmationTitle={t({
+              id: "chart-controls.delete.title",
+              message: "Delete chart?",
+            })}
+            confirmationText={t({
+              id: "chart-controls.delete.confirmation",
+              message: "Are you sure you want to delete this chart?",
+            })}
+            onClick={() => {
+              dispatch({ type: "CHART_CONFIG_REMOVE", value: { chartKey } });
+              handleClose();
+            }}
+            iconName="trash"
+            label={<Trans id="chart-controls.delete">Delete</Trans>}
+          />
+        ) : null}
+      </ArrowMenu>
+    </>
   );
 };
